Improve: Refactor alembic migration
