+++ .github/workflows/unittest.yml
@@ -46,7 +46,7 @@ jobs:
       - name: Install package
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install -r requirements.dev.txt
 
       - name: Set environment variables
         run: |
