--- .isort.cfg
@@ -1,5 +1,0 @@
-[settings]
-profile=black
-multi_line_output=3
-line_length=120
-include_trailing_comma=True
