--- alembic/versions/202410182112_9c33067267af_update_table_user.py
@@ -1,42 +1,0 @@
-"""Update table user
-
-Revision ID: 9c33067267af
-Revises: 1e42464a66e4
-Create Date: 2024-10-18 21:12:25.038651
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "9c33067267af"
-down_revision: Union[str, None] = "1e42464a66e4"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("users", sa.Column("first_name", sa.String(), nullable=True))
-    op.add_column("users", sa.Column("last_name", sa.String(), nullable=True))
-    op.add_column("users", sa.Column("full_name", sa.String(), nullable=True))
-    op.drop_constraint("users_email_key", "users", type_="unique")
-    op.drop_constraint("users_username_key", "users", type_="unique")
-    op.create_unique_constraint(op.f("UQ_users_email"), "users", ["email"])
-    op.create_unique_constraint(op.f("UQ_users_username"), "users", ["username"])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f("UQ_users_username"), "users", type_="unique")
-    op.drop_constraint(op.f("UQ_users_email"), "users", type_="unique")
-    op.create_unique_constraint("users_username_key", "users", ["username"])
-    op.create_unique_constraint("users_email_key", "users", ["email"])
-    op.drop_column("users", "full_name")
-    op.drop_column("users", "last_name")
-    op.drop_column("users", "first_name")
-    # ### end Alembic commands ###
