+++ alembic/versions/202410231652_c5355127adf1_create_user_table.py
@@ -1,0 +1,47 @@
+"""Create user table
+
+Revision ID: c5355127adf1
+Revises:
+Create Date: 2024-10-23 16:52:35.299860
+
+"""
+
+from typing import Sequence, Union
+
+import sqlalchemy as sa
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "c5355127adf1"
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "users",
+        sa.Column("username", sa.String(length=30), nullable=True),
+        sa.Column("first_name", sa.String(), nullable=True),
+        sa.Column("last_name", sa.String(), nullable=True),
+        sa.Column("full_name", sa.String(), nullable=True),
+        sa.Column("email", sa.String(length=50), nullable=True),
+        sa.Column("password", sa.String(length=200), nullable=True),
+        sa.Column("is_superuser", sa.Boolean(), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=True),
+        sa.Column("updated_at", sa.DateTime(), nullable=True),
+        sa.PrimaryKeyConstraint("id", name=op.f("PK_users")),
+        sa.UniqueConstraint("email", name=op.f("UQ_users_email")),
+        sa.UniqueConstraint("username", "email", name=op.f("UQ_users_username_email")),
+        sa.UniqueConstraint("username", name=op.f("UQ_users_username")),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("users")
+    # ### end Alembic commands ###
