+++ pyproject.toml
@@ -25,19 +25,27 @@ redis = "5.0.8"
 passlib = "1.7.4"
 psycopg2-binary = "2.9.9"
 pre-commit = "3.8.0"
+sqlalchemy-utils = "0.41.2"
+pyjwt = "2.9.0"
+email-validator = "^2.2.0"
+
+[tool.poetry.group.dev.dependencies]
 pytest = "8.3.2"
 pytest-cov = "5.0.0"
 pytest-env = "1.1.3"
+faker = "30.1.0"
 ruff = "0.6.4"
 flake8 = "7.1.1"
-sqlalchemy-utils = "0.41.2"
-faker = "30.1.0"
-pyjwt = "2.9.0"
-email-validator = "^2.2.0"
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.isort]
+profile = "black"
+multi_line_output = 3
+line_length = 120
+include_trailing_comma = true
 
 [tool.ruff]
 exclude = [
