+++ requirements.dev.txt
@@ -10,35 +10,52 @@ async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "4.0"
 asyncpg==0.29.0 ; python_version >= "3.10" and python_version < "4.0"
 attrs==24.2.0 ; python_version >= "3.10" and python_version < "4.0"
 cfgv==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
+colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
+coverage[toml]==7.6.4 ; python_version >= "3.10" and python_version < "4.0"
 distlib==0.3.9 ; python_version >= "3.10" and python_version < "4.0"
 dnspython==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
 email-validator==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
+exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
+faker==30.1.0 ; python_version >= "3.10" and python_version < "4.0"
 filelock==3.16.1 ; python_version >= "3.10" and python_version < "4.0"
+flake8==7.1.1 ; python_version >= "3.10" and python_version < "4.0"
 frozenlist==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
 greenlet==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
 gunicorn==23.0.0 ; python_version >= "3.10" and python_version < "4.0"
 identify==2.6.1 ; python_version >= "3.10" and python_version < "4.0"
 idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
+iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
 jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
 mako==1.3.6 ; python_version >= "3.10" and python_version < "4.0"
 markupsafe==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
+mccabe==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
 multidict==6.1.0 ; python_version >= "3.10" and python_version < "4.0"
 nodeenv==1.9.1 ; python_version >= "3.10" and python_version < "4.0"
 packaging==24.1 ; python_version >= "3.10" and python_version < "4.0"
 passlib==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
 platformdirs==4.3.6 ; python_version >= "3.10" and python_version < "4.0"
+pluggy==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
 pre-commit==3.8.0 ; python_version >= "3.10" and python_version < "4.0"
 propcache==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
 psycopg2-binary==2.9.9 ; python_version >= "3.10" and python_version < "4.0"
+pycodestyle==2.12.1 ; python_version >= "3.10" and python_version < "4.0"
 pydantic-core==2.23.2 ; python_version >= "3.10" and python_version < "4.0"
 pydantic==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
+pyflakes==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
 pyjwt==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
+pytest-cov==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
+pytest-env==1.1.3 ; python_version >= "3.10" and python_version < "4.0"
+pytest==8.3.2 ; python_version >= "3.10" and python_version < "4.0"
+python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
 python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
 pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
 redis==5.0.8 ; python_version >= "3.10" and python_version < "4.0"
+ruff==0.6.4 ; python_version >= "3.10" and python_version < "4.0"
+six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
 sqlalchemy-utils==0.41.2 ; python_version >= "3.10" and python_version < "4.0"
 sqlalchemy==2.0.34 ; python_version >= "3.10" and python_version < "4.0"
 swagger-ui-bundle==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
+tomli==2.0.2 ; python_version >= "3.10" and python_full_version <= "3.11.0a6"
 typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
 tzdata==2024.2 ; python_version >= "3.10" and python_version < "4.0"
 virtualenv==20.27.0 ; python_version >= "3.10" and python_version < "4.0"
